requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
PyAutoGUI==0.9.54
psutil==5.9.5
pywin32==306
newsapi-python==0.2.7
comtypes==1.2.0
python-dotenv==1.0.0
pycaw==20181226
PyAudio==0.2.13